requirements.txt
Django
django-ninja
rav
django-ninja-jwt[crypto]
pydantic[email]